requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.13.3
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
fastapi==0.115.0
fonttools==4.55.3
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.8
magic-filter==1.0.12
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib==3.10.0
multidict==6.1.0
numpy==2.2.1
packaging==24.2
pillow==11.1.0
propcache==0.2.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.20
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqladmin==0.20.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.31.0
WTForms==3.1.2
yarl==1.18.3
